Created Netbeans project, defined operation methods and ActionListener classes
Netbeans project:

The project has been created in NetBeans 11.1 as a standard Java Application.

Operation Methods:

The methods for processing mathematical operations have been written. They take two double values and returns the result as a double.

ActionListeners:

The handlers for the ActionListeners have been defined under the JavaCalculator.ActionHandlers folder in preparation for the GUI.