Updated ActionListener classes and created GUI class
ActionListener Classes:

Each ActionListener instance is now contained within an outer class for its particular group. The current outer classes are 'OpBtnHandlers' for the operation buttons and 'FileMenuHandlers' for the JMenuItems under the 'fileMenu' JMenu.

GUI Class:

There is now a class for creating the GUI used for this application. At the moment there is currently a JFrame,  JMenuBar (with a JMenu and JMenuItems) and JButtons. The JButtons and JMenuItems have their own ActionListeners.