Created JPanels for top and bottom halves, added Input area, Numpad and Operator buttons
Top and Bottom Halves:

To ensure the UI components have been laid out as intended, the Top and Bottom halves of the Container are now two separate JPanels, which will contain sub-panels for groups of UI components. As of this commit, the top half contains the input field and the bottom half contains the input buttons. The Container itself has been given a Y-Axis BoxLayout.

Input Area:

The top half of the UI now features a JLabel and JTextField for user input. These have been placed in a JPanel with a Y-Axis BoxLayout.

Numpad:

The number buttons have now been placed into a 3x4 GridLayout, in their own JPanel next to the Operator buttons.

Operator Buttons:

The operator buttons have been placed in a JPanel with a BoxLayout to the right of the numpad.