Added close event, and input validation for the TextField
Close Event:

When closing the main window, a confirmation dialog is shown, where the user can cancel if needed.

Input Validation:

Added a Document object to the JTextField, with an overridden insertString() that checks if the new character is numeric. If not, nothing happens, meaning there won't be any issues with invalid numbers down the line.